Create V.1.0
The purpose of the INCOCHAIN Routing Order is to utilize a decentralized platform and cryptocurrency to initiate smart contracts. The overall purpose of the Routing Order is a clear instruction depending on the incoterms or terms of sale for all international transport of goods and services. While this is the initiation of the routing order, and by paying with cryptocurrency, the key is to cut down on spam and junk routing orders. You have to pay to play. As you can see, there are several cryptocurrencies or blockchains available and each are capable while this is a use case for such applications as it relates to tokenization. Also, it is important to add that the INCOCHAIN Routing Order indicates method of payment, in cryptocurrency, never before seen in the world on a routing order. And also by way of example, this INCOCHAIN Routing Order offers time specificity as this example shows a 10% cashback that runs in conjunction with Byteball (GBYTE) 10% cashback program between buyer and seller, or in this case importer and exporter. In this regard, it could be an example of the competitive edge or advantage to secure new business development or to get 10% cashback, with a time lock making this example not only a routing order but time specific as to program, and not only instructs routing of shipment but accepted payment methods, not just hard currency. INCOCHAIN looks forward to leading the markets in the advancement of routing orders with the use of blockchain technology. This decentralizes things to the extent that ANYONE can initiate an INCOCHAIN Routing Order, not just industry specific players. Be on the look out for the sister to the INCOCHAIN Routing Order with the INCOCHAIN Sales Lead Generator that is used first traditionally between transportation or customhouse brokerage or trade finance companies, be it company owned offices or agent offices, to then get the routing order signed as instructed. With smart contracts, and with INCOCHAIN, anybody can initiate the INCOCHAIN Sales Lead Generator for whatever purpose they have, and such, payment in cryptocurrency will alleviate spam and junk sales leads.